fix tumblr feeds
This resolves two problems, identified with 'techist' tumblr feed but with broader ramifications.

Problem 1: feeds listed in reverse-chronological order

This is actually the most logical way to list an RSS feed, but the original code meant that if there was more than one
post older than the last updated date, empocketer would never update that date to the latest one, but rather to the *earliest*
(last) date in the list.

This commit adds a test in checker/checker.py to determine whether the current post in the loop is newer than the last published
value before it will update the database.

Problem 2: feeds with no 'updated', 'published' or 'modified' values

This is related to problem 1, but in combination makes both worse. Feeds without any of these values take the last_published value
from the last post in the feed. In reverse-chronologically ordered feeds, this is the *oldest* feed in the list, not the newest.
The commit adds a test to see which order the list is in, and uses the most recent date.