Fix renewed user token breaking app
Renaming and deleting feeds, and deleting user accounts, was tied to the user token that existed at the creation of the
feed or account. This breaks the app if the token changes for any reason: of which there could be several.

This update removes this constraint, checking against the username instead - in the case of feeds this means checking the
username against the parent list.

Fixes #3